requirements.txt
discord.py
feedparser
aiohttp
asyncio
python-dotenv